JB23AT2-32-Get-Hotel-room - adding controllers for Room and Reservations

+++ src/main/java/com/aacademy/moonlight/service/hotel/RoomService.java
@@ -2,16 +2,14 @@ package com.aacademy.moonlight.service.hotel;
 
 import com.aacademy.moonlight.dto.hotel.RoomRequest;
 import com.aacademy.moonlight.entity.hotel.Room;
-import com.aacademy.moonlight.entity.hotel.RoomType;
+
 
 import java.util.Optional;
 
 public interface RoomService {
 
     Room saveRoom(RoomRequest roomRequest);
-
-    Room findByNumber(Integer roomNumber);
+    Room findByNumber(Long roomNumber);
     Optional<Room> getRoom(Long id);
-
     void deleteRoom(Long id);
 }
